clean up and rename auth components
