--- app/assets/stylesheets/auth/sessions.scss
@@ -1,116 +1,0 @@
-// // Place all the styles related to the api/sessions controller here.
-// // They will automatically be included in application.css.
-// // You can use Sass (SCSS) here: http://sass-lang.com/
-
-// .nav-container {
-//   display: flex;
-//   justify-content: center;
-//   width: 100%;
-//   height: 100%;
-//   margin-top: 14px;
-//   margin-bottom: 14px;
-// }
-
-// #nav-general {
-//   // position: relative;
-//   display: flex;
-//   flex-direction: row;
-//   justify-content: space-between;
-//   align-items: center;
-//   // background-color: #FFDEDA;
-//   width: 1401px;
-//   min-width: 1401px;
-// }
-
-// #nav-categories-container-underline {
-//   border-bottom: 2px solid lightgray;
-// }
-
-// #nav-categories-container {
-//   display: flex;
-//   justify-content: center;
-//   // background-color: #FDB4C1;
-// }
-
-// #nav-cat-list {
-//   display: flex;
-//   justify-content: space-between;
-//   min-width: 1401px;
-//   width: 1401px;
-//   padding-top: .75%;
-//   padding-bottom: .65%;
-// }
-
-// .auth-container {
-//   display: flex;
-//   justify-content: space-between;
-//   align-items: center;
-//   width: 100%;
-// }
-
-// .form-title {
-//   font-size: 24px;
-// }
-
-// .login-nav {
-//   display: flex;
-//   justify-content: space-between;
-// }
-
-// .auth-container {
-//   display: block;
-//   width: 100%;
-// }
-
-// .auth-form-content, .auth-form-box {
-//   justify-content: space-evenly;
-//   display: flex;
-//   flex-direction: column;
-// }
-
-// .auth-form-content > input, .auth-form-box > input {
-//   height: 40px;
-// }
-
-// #auth-submit-button {
-//   background-color: #222222;
-//   color: white;
-//   width: 100%;
-//   border: 0;
-//   border-radius: 30px;
-//   height: 55px;
-//   margin-bottom: 20px;
-// }
-
-// #auth-submit-button:hover {
-//   background-color: #2f2f2f;
-// }
-
-// .user-auth-input {
-//   border-radius: 5px;
-//   width: 100%;
-// }
-
-// #login-button, #logout-button {
-//   position: relative;
-//   max-width: 70px;
-//   min-width: 70px;
-//   font-size: 18px;
-//   font-weight: bold;
-//   padding: 19%;
-//   padding-left: 3%;
-//   padding-right: 3%;
-// }
-
-// #signup-button{
-//   left: 259px;
-//   position: absolute;
-//   top: 21px;
-//   border-width: 2px;
-//   border-style: solid;
-//   padding: 2%;
-// }
-
-// #login-button:hover, #sign-up-button:hover, #logout-button:hover {
-//   background-color: lightgray;
-// }
